update rune test
