+++ packages/coin-bitcoin/package.json
@@ -23,22 +23,17 @@
   "devDependencies": {
     "@types/jest": "^27.5.1",
     "@types/node": "^12.20.52",
-    "bip32": "^4.0.0",
-    "bip39": "^3.1.0",
-    "bitcoinjs-lib": "^6.1.5",
-    "ecpair": "^2.1.0",
     "jest": "^27.5.1",
     "npm-run-all": "^4.1.5",
     "prettier": "2.6.2",
     "rimraf": "^3.0.2",
-    "tiny-secp256k1": "^2.2.3",
     "ts-jest": "^27.1.4",
     "ts-node": "^10.7.0",
     "typescript": "^4.6.2"
   },
   "dependencies": {
-    "@okxweb3/coin-base": "^1.0.0",
-    "@okxweb3/crypto-lib": "^1.0.0",
+    "@okxweb3/coin-base": "^1.0.5",
+    "@okxweb3/crypto-lib": "^1.0.1",
     "big-integer": "1.6.36"
   }
 }
